admin/index.tsx cleanup
