+++ src/program/utils/Seeds.tsx
@@ -1,0 +1,5 @@
+import * as buffer from "buffer";
+
+window.Buffer = buffer.Buffer;
+
+export const REGISTRY_SEEDS = Buffer.from("registry");
